+++ src/promptflow-devkit/tests/sdk_cli_test/e2etests/test_prompty_error_handler.py
@@ -296,12 +296,13 @@ class TestHandlePromptyError:
         assert mock_method.call_count == 1
         assert exc_info.value.error_codes == error_codes.split("/")
 
+    @pytest.mark.skipif(condition=not pytest.is_live, reason="OpenAI response failed.")
     @pytest.mark.parametrize(
         "max_tokens, error_message, error_codes, exception",
         [
             (0, "0 is less than the minimum of 1", "UserError/OpenAIError/BadRequestError", WrappedOpenAIError),
             (-1, "-1 is less than the minimum of 1", "UserError/OpenAIError/BadRequestError", WrappedOpenAIError),
-            ("invalid_max_token", "Unsupported data type", "UserError/OpenAIError/BadRequestError", WrappedOpenAIError),
+            ("invalid_max_token", "not of type 'integer'", "UserError/OpenAIError/BadRequestError", WrappedOpenAIError),
         ],
     )
     def test_invalid_max_tokens(self, azure_open_ai_connection, max_tokens, error_message, error_codes, exception):